Fixed subtle bug in libslpattr.c for 64-bit machines. unescape_into() takes a size_t * and is passed &val->unescape_len which was a int *. Fixed by changing the typedef in this header.

+++ openslp/libslpattr/libslpattr_internal.h
@@ -19,8 +19,8 @@
 
 typedef struct xx_value_t {
 	struct xx_value_t *next;
-	int escaped_len;
-	int unescaped_len;
+	size_t escaped_len;
+	size_t unescaped_len;
 	union {
 		SLPBoolean va_bool;
 		long va_int;
